app/requirements.txt
django==3.1
django-filter==2.4.0
djangorestframework==3.12
omdb==0.10
psycopg2-binary==2.8.5
